+++ views/roomView.ejs
@@ -48,6 +48,12 @@
     <nav class="navbar bg-body-tertiary">
         <div class="container">
             <a class="btn btn-primary" href="/f" role="button">Back</a>
+            <% if (typeof user !== 'undefined' && user) { %>
+                <h1>Welcome, <%= user.displayName %></h1>
+                <a href="/logout">Logout</a>
+            <% } else { %>
+                <a href="/auth/google">Login with Google</a>
+            <% } %>
         </div>
       </nav>
